Add files via upload
Decided to go the OpTab version B route. Will change the name to just OpTab.cpp/h at some point, but for now it's still OpTabB.cpp. Fully populated the opcode/mnemonic/format struct. Also uploaded a tester program and a makefile for compiling them. Still needs a little more debugging, error handling, possibly a bool method that just checks to see if the opcode is a valid opcode in the table, and also comments.