added a printRES method to dasm_main.cpp
printRES works smoothly with the SymLinkedList. It's contingent on all the RESW/RESB directives coming **after** the last instruction from the program counter. If he throws any RESW/RESB instructions in the middle of the object code, this isn't going to work.

also, edited SymLinkedList.h, SymLinkedList.cpp slightly, needed to move the Node to the header.